# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
...
        "title": "Waifu.im",
        "termsOfService": "https://www.waifu.im/terms-of-service",
        "contact": {
            "name": "Contact",
            "url": "http://www.waifu.im/contact",
            "email": "contact@waifu.im"
        },
        "license": {
            "name": "MPL 2.0",
            "url": "https://www.mozilla.org/en-US/MPL/2.0/"
        },
        "version": "5.2.2"
    },
    "host": "api.waifu.im",
    "basePath": "/",
    "paths": {
...
        "/fav/delete": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Removes an image from the user's favorites.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Favorites"
                ],
                "summary": "Delete an image from favorites",
                "parameters": [
                    {
                        "description": "Image data",
                        "name": "image",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serializers.Image"
                        }
                    },
                    {
                        "description": "User ID",
                        "name": "user_id",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/serializers.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "DELETED",
                        "schema": {
                            "$ref": "#/definitions/serializers.FavState"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/serializers.JSONError"
                        }
                    }
                }
            }
        },
        "/fav/insert": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Inserts an image to the user's favorites.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Favorites"
                ],
                "summary": "Insert an image to favorites",
                "parameters": [
                    {
                        "description": "Image data",
                        "name": "image",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serializers.Image"
                        }
                    },
                    {
                        "description": "User ID",
                        "name": "user_id",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/serializers.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "INSERTED",
                        "schema": {
                            "$ref": "#/definitions/serializers.FavState"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/serializers.JSONError"
                        }
                    }
                }
            }
        },
        "/fav/toggle": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Toggles an image in the user's favorites.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Favorites"
                ],
                "summary": "Toggle an image in favorites",
                "parameters": [
                    {
                        "description": "Image data",
                        "name": "image",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serializers.Image"
                        }
                    },
                    {
                        "description": "User ID",
                        "name": "user_id",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/serializers.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/serializers.FavState"
                        }
                    },
                    "default": {
                        "description": "",
                        "schema": {
                            "$ref": "#/definitions/serializers.JSONError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "serializers.FavState": {
            "type": "object",
            "properties": {
                "state": {
                    "type": "string",
                    "example": "INSERT"
                }
            }
        },
        "serializers.Image": {
            "type": "object",
            "properties": {
                "image_id": {
                    "type": "integer"
                }
            }
        },
...
        "serializers.JSONError": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                }
            }
        },
...
        "serializers.User": {
            "type": "object",
            "properties": {
                "user_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}